+++ src/main/java/com/mju/insuranceCompany/service/accident/exception/AccidentErrorCode.java
@@ -1,4 +1,4 @@
-package com.mju.insuranceCompany.global.exception.insurance;
+package com.mju.insuranceCompany.service.accident.exception;
 
 import com.mju.insuranceCompany.global.exception.ErrorCode;
 import lombok.Getter;
@@ -6,9 +6,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 
 @Getter @RequiredArgsConstructor
-public enum InsuranceErrorCode implements ErrorCode {
-
-    INSURANCE_ID_NOT_FOUND(HttpStatus.NOT_FOUND, "입력하신 ID에 해당하는 보험 정보가 존재하지 않습니다."),
+public enum AccidentErrorCode implements ErrorCode {
 
     ;
