+++ src/main/java/com/mju/insuranceCompany/service/user/exception/UnauthorizedAccessException.java
@@ -1,9 +1,8 @@
-package com.mju.insuranceCompany.global.exception.user;
+package com.mju.insuranceCompany.service.user.exception;
 
-import com.mju.insuranceCompany.global.exception.ErrorCode;
 import com.mju.insuranceCompany.global.exception.MyException;
 
-import static com.mju.insuranceCompany.global.exception.user.UserErrorCode.UNAUTHORIZED_ACCESS;
+import static com.mju.insuranceCompany.service.user.exception.UserErrorCode.UNAUTHORIZED_ACCESS;
 public class UnauthorizedAccessException extends MyException {
     public UnauthorizedAccessException() {
         super(UNAUTHORIZED_ACCESS);
